requirements.txt
# ============================================================================
# SAGE Requirements
# ============================================================================

# Core Machine Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.21.0

# Data Processing & Augmentation
albumentations>=1.3.0
pillow>=9.0.0
opencv-python>=4.5.0
scipy>=1.7.0
pandas>=1.5.0
scikit-learn>=1.0.0

# Configuration & I/O
pyyaml>=6.0
h5py>=3.7.0

# Visualization & Progress
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0

# Hyperparameter Search & Experiment Tracking (Optional/Merged)
ray[tune]>=2.9.0
mlflow>=2.10.0
hyperopt>=0.2.7
